Split bookmark tasks into separate files
